backend/requirements.txt
Django
djangorestframework
django-cors-headers
drf-spectacular
python-dotenv
pandas
reportlab
